+++ Changes.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.4] - 2026-03-09
 
 ### Added
 - Added high-level helpers for...
@@ -160,7 +160,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Original version.
 
-[Unreleased]: https://github.com/sanko/At.pm/compare/1.3...HEAD
+[Unreleased]: https://github.com/sanko/At.pm/compare/1.4...HEAD
+[1.4]: https://github.com/sanko/At.pm/compare/1.3...1.4
 [1.3]: https://github.com/sanko/At.pm/compare/1.2...1.3
 [1.2]: https://github.com/sanko/At.pm/compare/1.1...1.2
 [1.1]: https://github.com/sanko/At.pm/compare/1.0...1.1